+++ front/src/App.tsx
@@ -14,9 +14,9 @@ import DevPositionViewport from './View/DEV/Position/DevPositionViewport.test';
 import DevPositionPx from './View/DEV/Position/DevPositionPx.test';
 import Login from './View/Login/Login';
 import LoginIn from './View/Login/Login.in';
-import LoginId from './View/Login/Login.id';
-import LoginNickName from './View/Login/Login.nickname';
-import LoginPassWord from './View/Login/Login.password';
+import SignUpId from './View/Login/SignUp.id';
+import SignUpNickName from './View/Login/SignUp.nickname';
+import SignUpPassWord from './View/Login/SignUp.password';
 import SignUp from './View/Login/SignUp';
 import DevFetch from './View/DEV/DevFetch';
 
@@ -36,9 +36,9 @@ const App = () => {
           <Route path="/MyFactory" element={<MyFactory />} />
           <Route path="/Login" element={<Login />} />
           <Route path="/LoginIn" element={<LoginIn />} />
-          <Route path="/LoginId" element={<LoginId />} />
-          <Route path="/LoginNickName" element={<LoginNickName />} />
-          <Route path="/LoginPassWord" element={<LoginPassWord />} />
+          <Route path="/SignUpId" element={<SignUpId />} />
+          <Route path="/SignUpNickName" element={<SignUpNickName />} />
+          <Route path="/SignUpPassWord" element={<SignUpPassWord />} />
           <Route path="/SignUp" element={<SignUp />} />
           {/* DEVPosition */}
           <Route path="/Dev" element={<DevDesignSystem />} />
