+++ front/src/View/Login/SignUp.nickname.component.tsx
@@ -1,49 +1,27 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import loginBlackImage from '../DEV/img/login-black-logo.svg';
-import loginWhiteImage from '../DEV/img/login-white-logo.svg';
 import { useNavigate } from 'react-router-dom';
 
-const LoginId = () => {
-  const [isState, setState] = useState(false); // 메세지 없음
+const SignUpNickName = () => {
+  const [isState, setState] = useState(false);
+  const [isButton, setButton] = useState(false);
   const [isValue, setValue] = useState('');
-  const [isButton, setButton] = useState(false); // 회색
   const navigate = useNavigate();
-
-  // React.ChangeEvent<HTMLInputElement>
   const handleInput = (e: React.ChangeEvent<HTMLInputElement>) => {
     const value = e.target.value;
     setValue(value);
-    if (value !== '') {
-      // setState(!isState);
+    console.log(value.length);
+    if (value.length >= 6) {
+      setState(true);
+      setButton(false);
+    } else if (value.length === 0) {
       setState(false);
       setButton(false);
-      console.log(value);
-      if (isState === false) {
-        setButton(false);
-        setState(false);
-      }
-    } else {
+    } else if (value.length === 5) {
+      setState(false);
       setButton(true);
-      setState(true);
     }
   };
-
-  useEffect(() => {
-    console.log('useEffect test');
-    // if (isValue !== '') {
-    //   // setState(!isState);
-    //   setButton(false);
-    //   setState(false);
-    //   console.log(isValue);
-    //   if (isState === false) {
-    //     setButton(true);
-    //     setState(false);
-    //   }
-    // } else {
-    //   setButton(false);
-    //   setState(true);
-    // }
-  }, []);
 
   return (
     <div className="flex flex-col items-center justify-center w-full pt-[12.01%] pb-[clamp(0px,10.39%,56px)] bg-yellow-300">
@@ -57,12 +35,12 @@ const LoginId = () => {
       {/* 컨텐츠 영역 */}
       <div className="flex flex-col items-center justify-center w-full pt-[] pb-[] bg-yellow-100 mb-[clamp(0px,9.1%,69.45px)]">
         <div className="flex flex-col justify-start items-start w-full --Pretendard text-[33px] --bold mb-[clamp(0px,17.94%,136px)]">
-          <span>나만의 단어공장을</span>
-          <span>만들어볼까요?</span>
+          <span>마지막 단계에요!</span>
+          <span>공장을 위한 닉네임이 필요해요!</span>
         </div>
         <div className="flex flex-col w-full --Pretendard">
           <div className="--bold --font-m mb-[15px]">
-            공장을 위한 아이디가 필요해요!
+            공장을 위한 닉네임이 필요해요!
             <span className="--error-font-Color">*</span>
           </div>
           <div className="bg-white border-[1px] py-[3.69%] px-[3.96%] border-black rounded-[22px]">
@@ -76,34 +54,30 @@ const LoginId = () => {
           <div className="mt-[10px]">
             {isState ? (
               <span className="--error-font-Color --Pretendard --medium text-[14px]">
-                *올바르게 작성해주세요.
+                *너무 길어요! 올바르게 작성해주세요.
               </span>
             ) : (
-              <div className="--error-font-Color --Pretendard --medium text-[14px] mb-[24px]" />
+              <div className="--error-font-Color --Pretendard --medium text-[14px] mb-[20px]" />
             )}
           </div>
         </div>
       </div>
       <div className="flex flex-col items-center justify-center w-full">
         <div
-          className={`flex items-center justify-center w-full --Pretendard --semi-bold --font-xl ${isButton ? '--status-bg-Color-07' : '--status-bg-Color-01'} ${isButton ? '--status-font-Color-01' : '--status-font-Color-08'} py-[clamp(0px,3.3%,25px)] rounded-[30px] cursor-pointer`}
+          className={`flex items-center justify-center w-full --Pretendard --semi-bold --font-xl ${isButton ? '--primary-bg-Color' : '--status-bg-Color-07'} py-[clamp(0px,3.3%,25px)] rounded-[30px] ${isButton ? 'cursor-pointer' : ''}`}
           onClick={() => {
-            if (!isButton) {
-              console.log('navigate');
-              navigate('/LoginPassWord');
-            } else {
-              setState(true);
-            }
+            navigate('/LoginIn');
           }}
         >
           가입 하기
-          <img
-            className="ml-[16px]"
-            src={isButton ? loginBlackImage : loginWhiteImage}
-            alt="img"
-          />
+          <img className="ml-[16px]" src={loginBlackImage} alt="img" />
         </div>
-        <div className="--status-font-Color-04 border-[--status-font-Color-04] border-b-[1px] --Pretendard --medium --font-xs mt-[clamp(0px,1.98%,15px)] cursor-pointer">
+        <div
+          className="--status-font-Color-04 border-[--status-font-Color-04] border-b-[1px] --Pretendard --medium --font-xs mt-[clamp(0px,1.98%,15px)] cursor-pointer"
+          onClick={() => {
+            navigate('/LoginIn');
+          }}
+        >
           이미 만들어진 공장이 있어요!
         </div>
       </div>
@@ -111,4 +85,4 @@ const LoginId = () => {
   );
 };
 
-export default LoginId;
+export default SignUpNickName;
