+++ front/src/View/Login/SignUp.password.tsx
@@ -1,16 +1,16 @@
 import React from 'react';
 import Template from '../../global/components/Template.test';
-import LoginPassWordComponent from './Login.password.component';
-const LoginPassWord = () => {
+import SignUpPassWordComponent from './SignUp.password.component';
+const SignUpPassWord = () => {
   return (
     <div className="h-screen">
       <Template
         height={'h-[clamp(0px,91.12%,1149px)]'}
         shadowHeight={'h-[clamp(0px,91.12%,1149px)]'}
-        component={LoginPassWordComponent}
+        component={SignUpPassWordComponent}
       />
     </div>
   );
 };
 
-export default LoginPassWord;
+export default SignUpPassWord;
